+++ src/node-lib/curriculum-api-2023/queries/subjectListing/subjectListing.schema.ts
@@ -18,6 +18,7 @@ export const subjectSchema = z.object({
   pathwaySlug: pathwaySlugs.nullable(),
   pathwayTitle: pathways.nullable(),
   actions: actionsSchema,
+  features: z.object({}).passthrough().optional(),
 });
 
 const keyStageDataRaw = z.object({